refactor#60 camera position
